LayoutTests/fast/forms/select/optgroup-clicking-expected.txt

Test clicking with optgroup element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Click enabled option
PASS $("listbox").selectedIndex is 1

Click disabled option - doesn't change selectedIndex
PASS $("listbox").selectedIndex is 1

Menu list with arrow key
PASS $("menulist").selectedIndex is 8
PASS successfullyParsed is true

TEST COMPLETE
